[#19] DebugWindowBase 클래스에 일부 기능 위임
